% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{add_annotation_line_barplot_importance}
\alias{add_annotation_line_barplot_importance}
\title{Add a line bellow a gene from the function "importance_barplot"}
\usage{
add_annotation_line_barplot_importance(
  plot,
  permutation_df,
  last_gene,
  annotation = "most important features",
  adjust_annotation = 0.25,
  adjust_annotationX = 0.85
)
}
\arguments{
\item{plot}{ggplot object}

\item{permutation_df}{dataframe}

\item{last_gene}{str, name of gene bellow which to place the line}

\item{annotation}{str, to be displayed near the line}

\item{adjust_annotation}{float}

\item{adjust_annotationX}{float}
}
\description{
Add a line bellow a gene from the function "importance_barplot"
}
