% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{build_confusion_heatmap}
\alias{build_confusion_heatmap}
\title{Does what it says}
\usage{
build_confusion_heatmap(
  conf_mat,
  cell_order,
  filter = FALSE,
  column_filter = "signature_matrix",
  value_filter = "LM22",
  direction_color_palette = -1
)
}
\arguments{
\item{direction_color_palette}{}
}
\description{
Does what it says
}
