% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{compute_median_expr}
\alias{compute_median_expr}
\title{computes median returns a named vector}
\usage{
compute_median_expr(df, col_to_keep)
}
\arguments{
\item{df}{dataframe}

\item{col_to_keep}{str}
}
\description{
computes median returns a named vector
}
