% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{confidence_interval}
\alias{confidence_interval}
\title{computes confidence interval}
\usage{
confidence_interval(perm_df, n_perm = 20)
}
\arguments{
\item{perm_df}{dataframe}

\item{n_perm}{int}
}
\description{
computes confidence interval
}
