% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{confusion_matrix_setup}
\alias{confusion_matrix_setup}
\title{import files}
\usage{
confusion_matrix_setup(
  directory,
  pattern2,
  cells,
  signature_matrix,
  pattern1 = "confusion_heatmap_"
)
}
\arguments{
\item{pattern1}{}
}
\description{
import files
}
