% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{FC_prop_expr}
\alias{FC_prop_expr}
\title{Plot Fold change with proportion of expression}
\usage{
FC_prop_expr(df, n_cell, max_cell)
}
\arguments{
\item{df}{dataframe}

\item{max_cell}{}
}
\description{
Plot Fold change with proportion of expression
}
