% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{geneset_to_binary_df}
\alias{geneset_to_binary_df}
\title{takes a list of geneset, returns a dataframe with 0 and 1, genes in row, cell in column}
\usage{
geneset_to_binary_df(geneset_list)
}
\arguments{
\item{geneset_list}{list}
}
\description{
takes a list of geneset, returns a dataframe with 0 and 1, genes in row, cell in column
}
