% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_genes_from_binary}
\alias{get_genes_from_binary}
\title{extract genes from binary dataframe. Does the inverse of geneset_to_binary_df}
\usage{
get_genes_from_binary(
  binary_df,
  expressed_genes = NULL,
  filter_expressed = FALSE
)
}
\arguments{
\item{binary_df}{dataframe with 0 and 1}

\item{expressed_genes}{vector set of genes to keep}

\item{filter_expressed}{FALSE}
}
\description{
extract genes from binary dataframe. Does the inverse of geneset_to_binary_df
}
