% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_permutation_list}
\alias{get_permutation_list}
\title{Imports csv}
\usage{
get_permutation_list(directory, pattern2, pattern1 = "Permutation_result_")
}
\arguments{
\item{pattern1}{}
}
\description{
Imports csv
}
