% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{heatmap_from_geneset_list}
\alias{heatmap_from_geneset_list}
\title{Heatmap for geneset list. red if genes in geneset, darkgrey if not}
\usage{
heatmap_from_geneset_list(
  geneset_list,
  tabula_column,
  title = " ",
  min_representation = 2,
  cutree_rows = 1,
  clust = TRUE
)
}
\arguments{
\item{geneset_list}{list}

\item{tabula_column}{str, name of column from tabula sapiens}

\item{title}{str}

\item{min_representation}{int display genes represented a min number of times}

\item{cutree_rows}{int}

\item{clust}{Boolean}
}
\description{
Heatmap for geneset list. red if genes in geneset, darkgrey if not
}
