% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{importance_barplot}
\alias{importance_barplot}
\title{Build the importance barlot}
\usage{
importance_barplot(
  permutation_df,
  permutation_number = 50,
  filter = "mean",
  title = "Mean feature importance barplot",
  conf_interval = TRUE
)
}
\arguments{
\item{permutation_df}{dataframe}

\item{permutation_number}{int}

\item{filter}{type of filtration: among c(bsup,binf,zero,mean)
bsup : greater threshold of the mean confidence interval
binf : lower threshold of the mean confidence interval
zero : remove mean importance value if imp = 0 or inferior
mean : keeps genes with importance above mean}

\item{title}{str, title of plot}

\item{conf_interval}{Boolean}
}
\description{
Build the importance barlot
}
