% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{kneedle_threshold}
\alias{kneedle_threshold}
\title{Computes threshold with kneedle algorithm}
\usage{
kneedle_threshold(df, y, sens = 1, min = 0)
}
\arguments{
\item{df}{dataframe}

\item{y}{vector}

\item{sens}{int}

\item{min}{int}
}
\description{
Computes threshold with kneedle algorithm
}
