% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{merge_confusion_matrix}
\alias{merge_confusion_matrix}
\title{Does what it says}
\usage{
merge_confusion_matrix(pattern2, tbs_path, lm22_path, til10_path, cells)
}
\arguments{
\item{cells}{}
}
\description{
Does what it says
}
