% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{RA_plot}
\alias{RA_plot}
\title{Ratio - Average Plot}
\usage{
RA_plot(
  FC_df,
  x,
  y,
  title = "RA plot",
  highlight = c("TOP20"),
  annotation = "Most Important Features",
  vjust = 2,
  force = 3,
  y_scale = 0,
  add_label = TRUE
)
}
\arguments{
\item{FC_df}{dataframe}

\item{x}{str, name of the column to be displayed in x}

\item{y}{str, name of the column to be displayed in x}

\item{title}{str,}

\item{highlight}{str, type of highlighting: 'TOP20' to highlight the top20 genes with highest FC or a specific gene to precise}

\item{annotation}{str to be displayed on graĥ}

\item{vjust}{int}

\item{force}{int}

\item{y_scale}{int}

\item{add_label}{Boolean}
}
\description{
Ratio - Average Plot
}
