% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{Rprop_plot}
\alias{Rprop_plot}
\title{Plot Proportion of expre}
\usage{
Rprop_plot(
  FC_df,
  FC_x,
  y,
  title = "plot",
  highlight = c("TOP20"),
  annotation = "Most Important Features",
  vjust = 2,
  force = 3,
  x_scale = 0,
  add_label = TRUE,
  FC_threshold = 1,
  prop_threshold = 0.25,
  FC_thresh2 = 1,
  prop_threshold_to_show = prop_threshold
)
}
\arguments{
\item{FC_df}{dataframe}

\item{FC_x}{Fold change}

\item{y}{vector}

\item{title}{str}

\item{highlight}{str}

\item{annotation}{str}

\item{vjust}{int}

\item{force}{int}

\item{x_scale}{int}

\item{add_label}{Boolean}

\item{FC_threshold}{int}

\item{prop_threshold}{int}

\item{FC_thresh2}{int}

\item{prop_threshold_to_show}{int}
}
\description{
Plot Proportion of expre
}
