% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{visualize_annotation}
\alias{visualize_annotation}
\title{Query GO annotation for a set of gene and returns a visualisation}
\usage{
visualize_annotation(Annotation_table, celltype)
}
\arguments{
\item{Annotation_table}{dataframe}

\item{celltype}{str, in ttitle of the graph}
}
\description{
Query GO annotation for a set of gene and returns a visualisation
}
